🏗️ structure: move structure of Popover main component
