+++ packages/popover/src/components/PopoverFooter.tsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import { PopoverProps } from '../popoverTypes';
+import { PopoverProps } from '../types';
 
 export const popoverFooter = (props: PopoverProps) => {
   return <div>{props.children}</div>;
